+++ dynamic-amm-quote/src/curve/mod.rs
@@ -1,2 +1,1 @@
-// pub mod curve_type;
-pub mod fees;
+pub mod curve_type;