+++ programs/dynamic-amm/src/constants.rs
@@ -8,6 +8,8 @@ pub static MIN_CHANGE_AMP_DURATION: u64 = 600; // 10 minutes
 pub mod seeds {
     /// Prefix for apy PDA
     pub static APY_PREFIX: &str = "apy";
+    /// Prefix for config PDA
+    pub static CONFIG_PREFIX: &[u8] = b"config";
 }
 
 /// Store constants related to fees
@@ -20,13 +22,13 @@ pub mod fee {
     // 1bps, https://curve.fi/rootfaq
     pub static STABLE_SWAP_TRADE_FEE_NUMERATOR: u64 = 10;
 
-    /// Admin trade fee numerator for constant product swap curve.
+    /// Protocol trade fee numerator for constant product swap curve.
     // 5bps, https://docs.uniswap.org/protocol/V2/concepts/advanced-topics/fees
-    pub static CONSTANT_PRODUCT_ADMIN_TRADE_FEE_NUMERATOR: u64 = 50;
+    pub static CONSTANT_PRODUCT_PROTOCOL_TRADE_FEE_NUMERATOR: u64 = 50;
 
-    /// Admin trade fee numerator for stable swap curve.
+    /// Protocol trade fee numerator for stable swap curve.
     // 2bps, https://curve.fi/rootfaq
-    pub static STABLE_SWAP_ADMIN_TRADE_FEE_NUMERATOR: u64 = 5;
+    pub static STABLE_SWAP_PROTOCOL_TRADE_FEE_NUMERATOR: u64 = 5;
 
     /// Host trade fee numerator
     // 20% of admin trade fee