+++ ts-client/src/amm/tests/utils/index.ts
@@ -18,6 +18,13 @@ export const airDropSol = async (connection: Connection, publicKey: PublicKey, a
   }
 };
 
+export const airDropSolIfBalanceNotEnough = async (connection: Connection, publicKey: PublicKey, balance = 1) => {
+  const walletBalance = await connection.getBalance(publicKey);
+  if (walletBalance < balance * LAMPORTS_PER_SOL) {
+    await airDropSol(connection, publicKey);
+  }
+};
+
 export const getOrCreateATA = async (connection: Connection, mint: PublicKey, owner: PublicKey, payer: Keypair) => {
   const token = new Token(connection, mint, TOKEN_PROGRAM_ID, payer);
   const ata = await token.getOrCreateAssociatedAccountInfo(owner);